auth_server/requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
Flask==2.3.2
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
Werkzeug==2.3.4